requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
pydantic
python-dotenv
cryptography
python-jose[cryptography]
passlib[bcrypt]
python-multipart